Type RedTrack swing function slots

+++ src/RedSound/RedMidiCtrl.cpp
@@ -1822,7 +1822,7 @@ static void __MidiCtrl_VibrateOn(RedSoundCONTROL* control, RedKeyOnDATA* keyOn, 
 
     divisor = depth;
     track->m_vibrateRate = REDSOUND_MIDI_RATE_FIXED_NUMERATOR / divisor;
-    track->m_vibrateFunc = (int)SwingEntryFunction[track->m_command[2] & REDSOUND_MIDI_SWING_FUNC_MASK];
+    track->m_vibrateFunc = SwingEntryFunction[track->m_command[2] & REDSOUND_MIDI_SWING_FUNC_MASK];
     track->m_vibrateRateDelta = track->m_vibrateDepthDelta = 0;
     track->m_command += 3;
 
@@ -1966,7 +1966,7 @@ static void __MidiCtrl_VibrateType(RedSoundCONTROL*, RedKeyOnDATA*, RedTrackDATA
 {
 	u32 type = *track->m_command++;
 
-	track->m_vibrateFunc = (u32)SwingEntryFunction[type & REDSOUND_MIDI_SWING_FUNC_MASK];
+	track->m_vibrateFunc = SwingEntryFunction[type & REDSOUND_MIDI_SWING_FUNC_MASK];
 }
 
 /*
@@ -2009,7 +2009,7 @@ static void __MidiCtrl_TremoloOn(RedSoundCONTROL*, RedKeyOnDATA*, RedTrackDATA* 
 	}
 	divisor = rateDivisor;
 	track->m_tremoloRate = REDSOUND_MIDI_RATE_FIXED_NUMERATOR / divisor;
-	track->m_tremoloFunc = (int)SwingEntryFunction[track->m_command[2] & REDSOUND_MIDI_SWING_FUNC_MASK];
+	track->m_tremoloFunc = SwingEntryFunction[track->m_command[2] & REDSOUND_MIDI_SWING_FUNC_MASK];
 	track->m_tremoloRateDelta = track->m_tremoloDepthDelta = 0;
 	track->m_command += 3;
 
@@ -2155,7 +2155,7 @@ static void __MidiCtrl_TremoloType(RedSoundCONTROL*, RedKeyOnDATA*, RedTrackDATA
 {
 	u32 type = *track->m_command++;
 
-	track->m_tremoloFunc = (u32)SwingEntryFunction[type & REDSOUND_MIDI_SWING_FUNC_MASK];
+	track->m_tremoloFunc = SwingEntryFunction[type & REDSOUND_MIDI_SWING_FUNC_MASK];
 }
 
 /*
@@ -2196,7 +2196,7 @@ static void __MidiCtrl_ShakeOn(RedSoundCONTROL*, RedKeyOnDATA*, RedTrackDATA* tr
 	}
 	divisor = rate;
 	track->m_shakeRate = REDSOUND_MIDI_RATE_FIXED_NUMERATOR / divisor;
-	track->m_shakeFunc = (int)SwingEntryFunction[track->m_command[2] & REDSOUND_MIDI_SWING_FUNC_MASK];
+	track->m_shakeFunc = SwingEntryFunction[track->m_command[2] & REDSOUND_MIDI_SWING_FUNC_MASK];
 	track->m_shakeRateDelta = track->m_shakeDepthDelta = 0;
 	track->m_shakeOutput = 0;
 	track->m_shakePan = 0;
@@ -2320,7 +2320,7 @@ static void __MidiCtrl_ShakeType(RedSoundCONTROL*, RedKeyOnDATA*, RedTrackDATA* 
 {
 	u32 type = *track->m_command++;
 
-	track->m_shakeFunc = (u32)SwingEntryFunction[type & REDSOUND_MIDI_SWING_FUNC_MASK];
+	track->m_shakeFunc = SwingEntryFunction[type & REDSOUND_MIDI_SWING_FUNC_MASK];
 }
 
 /*
